feat: build frontend in backend workflow
This commit modifies the backend-macos-arm64.yml workflow to build the frontend before building the backend. This ensures that the ttyd binary always has the latest version of the frontend code.

This commit also adds src/html.h to .gitignore and removes it from the repository, as this file is generated during the build process and should not be versioned.